requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
openai>=1.35.3
faiss-cpu>=1.8.0
numpy>=1.26.4
chromadb>=0.5.3
